Commit branch master n°2
-Modif of manifest
-adding new classes for the game structure: the game object classes which basically represent
the object that will be used in game
- ->ball class
    - define ball class structure
    - define the ball class moving behavior(work well)
    - basic tests

+++ src/GameTool/Graphics.java
@@ -7,20 +7,20 @@ import android.graphics.Paint;
 public interface Graphics 
 {
 //getters
- public int getX();
- public int getY();
+ public float getX();
+ public float getY();
  
 
  //setters
- public void setX(int X);
- public void setY(int Y);
+ public void setX(float X);
+ public void setY(float Y);
  public void setPaint(Paint p);
  
  //add/sub methods
- public void addX(int X);
- public void addY(int Y);
- public void subX(int X);
- public void subY(int Y);
+ public void addX(float X);
+ public void addY(float Y);
+ public void subX(float X);
+ public void subY(float Y);
  //
  public void draw(Canvas canvas);
